Nyay-AI-main/requirements.txt
streamlit>=1.24.0
python-dotenv>=0.21.0
folium>=0.14.0
streamlit-folium>=0.13.0
geopy>=2.3.0
requests>=2.31.0
crewai>=0.11.0
langchain>=0.1.0
